% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKNN.R
\name{beefedUpKNN}
\alias{beefedUpKNN}
\title{fSVA and k-Nearest Neighbor classifier}
\usage{
beefedUpKNN(x, y, n.sv, k, ...)
}
\arguments{
\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{y}{factor: Class labels.}

\item{n.sv}{integer: Number of surrogate variables to estimate.}

\item{k}{integer: Number of nearest neighbors casting votes.}

\item{...}{additional arguments passed to smartsva.}
}
\value{
beefedUpKNN object.
}
\description{
Couples together fSVA with Nearest Shrunken Centroid classifier.
}
\examples{
# ADD_EXAMPLES_HERE
}
