% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSNC.R
\name{beefedUpSNC}
\alias{beefedUpSNC}
\title{fSVA and Nearest Shrunken Centroid classifier}
\usage{
beefedUpSNC(x, y, n.sv, threshold, ...)
}
\arguments{
\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{y}{factor: Class labels.}

\item{n.sv}{integer: Number of surrogate variables to estimate.}

\item{threshold}{numeric: Amount of centroid shrinkage (controls how many features are retained in the model).}

\item{...}{additional arguments passed to smartsva.}
}
\value{
beefedUpNSC object.
}
\description{
Couples together fSVA with Nearest Shrunken Centroid classifier.
}
\examples{
# ADD_EXAMPLES_HERE
}
