% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{caretKNN}
\alias{caretKNN}
\title{fSVA and k Nearest Neighbor caret-package custom model.}
\format{List of model components required by caret.}
\usage{
caretKNN
}
\description{
Wrapper for using fSVA and Nearest Shrunken Centroid classifier within the caret framework.
}
\keyword{datasets}
