% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSNC.R
\name{featureImportance}
\alias{featureImportance}
\title{fSVA and Nearest Shrunken Centroid feature importance}
\usage{
featureImportance(object, threshold = NULL)
}
\arguments{
\item{object}{beefedUpSNC object.}

\item{threshold}{numeric: Amount of centroid shrinkage (controls how many features are retained in the model).}
}
\value{
data.frame.
}
\description{
DETAILS.
}
\examples{
# ADD_EXAMPLES_HERE
}
