% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSVA.R
\name{fitSVA}
\alias{fitSVA}
\title{Surrogate Variable Analysis (SVA) wrapper for prediction}
\usage{
fitSVA(x, y, n.sv, ...)
}
\arguments{
\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{y}{factor: Class labels.}

\item{n.sv}{integer: Number of surrogate variables to estimate.}

\item{...}{additional arguments passed to smartsva.}
}
\value{
fitSVA object (With the following components...)
}
\description{
Wraps the smartsva and fsva functions into a standard set of fit and predict functions.
}
\examples{
# ADD_EXAMPLES_HERE
}
