% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn4.R
\name{knn4}
\alias{knn4}
\title{k-Nearest Neighbor classifier with multiple k}
\usage{
knn4(x, y, k)
}
\arguments{
\item{x}{a matrix of training set predictors}

\item{y}{a factor vector of training set classes}

\item{k}{number of neighbours considered.}
}
\value{
An object of class \code{knn3}. See \code{\link{predict.knn3}}.
}
\description{
k-Nearest Neighbor classifier with multiple k
}
\details{
\code{knn3} is essentially the same code as \code{\link[ipred]{ipredknn}}
and \code{knn3Train} is a copy of \code{\link[class]{knn}}. The underlying C
code from the \code{class} package has been modified to return the vote
percentages for each class (previously the percentage for the winning class
was returned).
}
