% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSNC.R
\name{plot.beefedUpSNC}
\alias{plot.beefedUpSNC}
\title{Plot batch-corrected centroids}
\usage{
\method{plot}{beefedUpSNC}(x, threshold = NULL, ...)
}
\arguments{
\item{x}{beefedUpSNC object.}

\item{threshold}{numeric: Amount of centroid shrinkage (controls how many features are retained in the model).}

\item{...}{additional arguments not currently used.}
}
\value{
ggplot2 object.
}
\description{
Plots the results of a fSVA and Nearest Shrunken Centroid analysis as a PCA-plot.
}
\examples{
# ADD_EXAMPLES_HERE
}
