% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKNN.R
\name{predict.beefedUpKNN}
\alias{predict.beefedUpKNN}
\title{fSVA and k-Nearest Neighbor prediction}
\usage{
\method{predict}{beefedUpKNN}(object, x, k = NULL, type = "class", ...)
}
\arguments{
\item{object}{beefedUpKNN object.}

\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{k}{integer: Number of nearest neighbors casting votes.}

\item{type}{character: Either "class" for labels or "prob" for class probabilities.}

\item{...}{additional arguments, currently not used.}
}
\value{
factor if type = "class" or matrix of class probabilities if type = "posterior"
}
\description{
Predict labels or probabilities for new samples.
}
\examples{
# ADD_EXAMPLES_HERE
}
