% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSNC.R
\name{predict.beefedUpSNC}
\alias{predict.beefedUpSNC}
\title{fSVA and Nearest Shrunken Centroid prediction}
\usage{
\method{predict}{beefedUpSNC}(object, x, threshold = NULL, type = "class",
  ...)
}
\arguments{
\item{object}{beefedUpSNC object.}

\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{threshold}{numeric: Amount of centroid shrinkage (controls how many features are retained in the model).}

\item{type}{character: Either "class" for labels or "posterior" for class probabilities.}

\item{...}{additional arguments, currently not used.}
}
\value{
factor if type = "class" or matrix of class probabilities if type = "posterior"
}
\description{
Predict labels or posterior probabilities for new samples.
}
\examples{
# ADD_EXAMPLES_HERE
}
