% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSVA.R
\name{predict.fitSVA}
\alias{predict.fitSVA}
\title{Frozen SVA (fSVA) wrapper for prediction}
\usage{
\method{predict}{fitSVA}(object, x, ...)
}
\arguments{
\item{object}{fitSVA object.}

\item{x}{matrix: Normalized and log-transformed expression values (samples in rows, features in columns).}

\item{...}{additional arguments, not currently used.}
}
\value{
matrix with batch corrected expression for the new samples.
}
\description{
Given a SVA, batch correct new samples.
}
\examples{
# ADD_EXAMPLES_HERE
}
